projects/dibt/synthetic_evaluation/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
aiohttp==3.9.5
    # via
    #   datasets
    #   fsspec
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.6.0
    # via pydantic
anyio==4.3.0
    # via httpx
attrs==23.2.0
    # via aiohttp
certifi==2024.2.2
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via typer
datasets==2.19.0
    # via distilabel
dill==0.3.8
    # via
    #   datasets
    #   multiprocess
distilabel @ git+https://github.com/argilla-io/distilabel.git
    # via -r requirements.in
filelock==3.13.4
    # via
    #   datasets
    #   huggingface-hub
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
fsspec[http]==2024.3.1
    # via
    #   datasets
    #   huggingface-hub
h11==0.14.0
    # via httpcore
httpcore==1.0.5
    # via httpx
httpx==0.27.0
    # via distilabel
huggingface-hub==0.22.2
    # via datasets
idna==3.7
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
jinja2==3.1.3
    # via distilabel
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.5
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
multiprocess==0.70.16
    # via
    #   datasets
    #   distilabel
nest-asyncio==1.6.0
    # via distilabel
networkx==3.3
    # via distilabel
numpy==1.26.4
    # via
    #   datasets
    #   pandas
    #   pyarrow
    #   scipy
packaging==24.0
    # via
    #   datasets
    #   huggingface-hub
pandas==2.2.2
    # via datasets
pyarrow==16.0.0
    # via datasets
pyarrow-hotfix==0.6
    # via datasets
pydantic==2.7.1
    # via distilabel
pydantic-core==2.18.2
    # via pydantic
pygments==2.17.2
    # via rich
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.0.1
    # via -r requirements.in
pytz==2024.1
    # via pandas
pyyaml==6.0.1
    # via
    #   datasets
    #   huggingface-hub
requests==2.31.0
    # via
    #   datasets
    #   huggingface-hub
rich==13.7.1
    # via
    #   distilabel
    #   typer
scipy==1.13.0
    # via distilabel
shellingham==1.5.4
    # via typer
six==1.16.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
tblib==3.0.0
    # via distilabel
tqdm==4.66.2
    # via
    #   datasets
    #   huggingface-hub
typer==0.12.3
    # via distilabel
typing-extensions==4.11.0
    # via
    #   huggingface-hub
    #   pydantic
    #   pydantic-core
    #   typer
tzdata==2024.1
    # via pandas
urllib3==2.2.1
    # via requests
xxhash==3.4.1
    # via datasets
yarl==1.9.4
    # via aiohttp
